requirements.txt
Flask
scikit-surprise
pandas